+++ lib/widgets/divider.dart
@@ -1,18 +1,24 @@
 part of kh_easy_dev;
 
-Widget kheasydevDivider({double? height, EdgeInsets? padding}) {
+Widget kheasydevDivider({double? height, EdgeInsets? padding, bool? black}) {
   return Padding(
     padding: padding ?? const EdgeInsets.only(top: 8),
     child: SizedBox(
       height: height ?? 1,
       child: Container(
-        decoration: const BoxDecoration(
+        decoration: BoxDecoration(
           gradient: LinearGradient(
-            colors: [
-              Colors.white10,
-              Colors.white,
-              Colors.white10,
-            ],
+            colors: black == true
+                ? [
+                    Colors.black12,
+                    Colors.black,
+                    Colors.black12,
+                  ]
+                : [
+                    Colors.white10,
+                    Colors.white,
+                    Colors.white10,
+                  ],
             stops: [0.0, 0.5, 1.0],
             begin: Alignment.centerRight,
             end: Alignment.centerLeft,
@@ -23,20 +29,27 @@ Widget kheasydevDivider({double? height, EdgeInsets? padding}) {
   );
 }
 
-Widget kheasydevVerticalDivider({double? height, EdgeInsets? padding}) {
+Widget kheasydevVerticalDivider(
+    {double? height, EdgeInsets? padding, bool? black}) {
   return Padding(
     padding: const EdgeInsets.symmetric(horizontal: 16),
     child: SizedBox(
       width: 1,
       height: 60,
       child: Container(
-        decoration: const BoxDecoration(
+        decoration: BoxDecoration(
           gradient: LinearGradient(
-            colors: [
-              Colors.white10,
-              Colors.white,
-              Colors.white10,
-            ],
+            colors: black == true
+                ? [
+                    Colors.black12,
+                    Colors.black,
+                    Colors.black12,
+                  ]
+                : [
+                    Colors.white10,
+                    Colors.white,
+                    Colors.white10,
+                  ],
             stops: [0.0, 0.5, 1.0],
             begin: Alignment.topCenter,
             end: Alignment.bottomCenter,